Moving reducers to separate files
